[HDM-28] searchbar component refactor / hover/active state fix
