--- packages/apps/dashboard/ui-2024/src/components/Search/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Search';
