+++ packages/apps/dashboard/ui-2024/src/theme.tsx
@@ -272,6 +272,15 @@ const theme: ThemeOptions = createTheme({
         },
       },
     },
+    MuiMenuItem: {
+      styleOverrides: {
+        root: {
+          '&:hover': {
+            backgroundColor: '#1406B207)',
+          },
+        },
+      },
+    },
   },
 });
